Move API conversion tests into sub-module
This patch relocates the API conversion tests into a distinct
sub-module under the existing API module. This reduces the API
module's dependency graph.
