+++ api/test/utilconversion/conversion.go
@@ -1,0 +1,68 @@
+/*
+Copyright 2019 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// NOTE: These functions are borrowed from upstream cluster-api repo.
+
+package utilconversion
+
+import (
+	"encoding/json"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+// AnnotationKey is the annotation key to ensure lossless
+// conversion of objects between API versions.
+const AnnotationKey = "vmoperator.vmware.com/conversion"
+
+// MarshalData stores the source object as json data in the destination object annotations map.
+// It ignores the metadata of the source object.
+func MarshalData(src metav1.Object, dst metav1.Object) error {
+	u, err := runtime.DefaultUnstructuredConverter.ToUnstructured(src)
+	if err != nil {
+		return err
+	}
+	delete(u, "metadata")
+
+	data, err := json.Marshal(u)
+	if err != nil {
+		return err
+	}
+	annotations := dst.GetAnnotations()
+	if annotations == nil {
+		annotations = map[string]string{}
+	}
+	annotations[AnnotationKey] = string(data)
+	dst.SetAnnotations(annotations)
+	return nil
+}
+
+// UnmarshalData tries to retrieve the data from the annotation and
+// unmarshals it into the object passed as input.
+func UnmarshalData(from metav1.Object, to interface{}) (bool, error) {
+	annotations := from.GetAnnotations()
+	data, ok := annotations[AnnotationKey]
+	if !ok {
+		return false, nil
+	}
+	if err := json.Unmarshal([]byte(data), to); err != nil {
+		return false, err
+	}
+	delete(annotations, AnnotationKey)
+	from.SetAnnotations(annotations)
+	return true, nil
+}
