+++ api/test/v1alpha2/conversion_test.go
@@ -2,7 +2,7 @@
 // The term “Broadcom” refers to Broadcom Inc. and/or its subsidiaries.
 // SPDX-License-Identifier: Apache-2.0
 
-package v1alpha3_test
+package v1alpha2_test
 
 import (
 	"encoding/json"
@@ -16,8 +16,8 @@ import (
 	runtimeserializer "k8s.io/apimachinery/pkg/runtime/serializer"
 	"sigs.k8s.io/randfill"
 
-	"github.com/vmware-tanzu/vm-operator/api/utilconversion/fuzztests"
-	vmopv1a3 "github.com/vmware-tanzu/vm-operator/api/v1alpha3"
+	"github.com/vmware-tanzu/vm-operator/api/test/utilconversion/fuzztests"
+	vmopv1a2 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha5"
 	vmopv1sysprep "github.com/vmware-tanzu/vm-operator/api/v1alpha5/sysprep"
 )
@@ -31,7 +31,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 
 	BeforeEach(func() {
 		scheme = runtime.NewScheme()
-		Expect(vmopv1a3.AddToScheme(scheme)).To(Succeed())
+		Expect(vmopv1a2.AddToScheme(scheme)).To(Succeed())
 		Expect(vmopv1.AddToScheme(scheme)).To(Succeed())
 	})
 
@@ -44,7 +44,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachine{},
-				Spoke:  &vmopv1a3.VirtualMachine{},
+				Spoke:  &vmopv1a2.VirtualMachine{},
 				FuzzerFuncs: []fuzzer.FuzzerFuncs{
 					overrideVirtualMachineFieldsFuncs,
 				},
@@ -67,7 +67,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineClass{},
-				Spoke:  &vmopv1a3.VirtualMachineClass{},
+				Spoke:  &vmopv1a2.VirtualMachineClass{},
 			}
 		})
 		Context("Spoke-Hub-Spoke", func() {
@@ -87,7 +87,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineImage{},
-				Spoke:  &vmopv1a3.VirtualMachineImage{},
+				Spoke:  &vmopv1a2.VirtualMachineImage{},
 				FuzzerFuncs: []fuzzer.FuzzerFuncs{
 					overrideVirtualMachineImageFieldsFuncs,
 				},
@@ -110,32 +110,9 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.ClusterVirtualMachineImage{},
-				Spoke:  &vmopv1a3.ClusterVirtualMachineImage{},
+				Spoke:  &vmopv1a2.ClusterVirtualMachineImage{},
 				FuzzerFuncs: []fuzzer.FuzzerFuncs{
 					overrideVirtualMachineImageFieldsFuncs,
-				},
-			}
-		})
-		Context("Spoke-Hub-Spoke", func() {
-			It("should get fuzzy with it", func() {
-				fuzztests.SpokeHubSpoke(input)
-			})
-		})
-		Context("Hub-Spoke-Hub", func() {
-			It("should get fuzzy with it", func() {
-				fuzztests.HubSpokeHub(input)
-			})
-		})
-	})
-
-	Context("VirtualMachineImageCache", func() {
-		BeforeEach(func() {
-			input = fuzztests.FuzzTestFuncInput{
-				Scheme: scheme,
-				Hub:    &vmopv1.VirtualMachineImageCache{},
-				Spoke:  &vmopv1a3.VirtualMachineImageCache{},
-				FuzzerFuncs: []fuzzer.FuzzerFuncs{
-					overrideVirtualMachineImageCacheFieldsFuncs,
 				},
 			}
 		})
@@ -156,7 +133,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachinePublishRequest{},
-				Spoke:  &vmopv1a3.VirtualMachinePublishRequest{},
+				Spoke:  &vmopv1a2.VirtualMachinePublishRequest{},
 			}
 		})
 		Context("Spoke-Hub-Spoke", func() {
@@ -176,7 +153,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineService{},
-				Spoke:  &vmopv1a3.VirtualMachineService{},
+				Spoke:  &vmopv1a2.VirtualMachineService{},
 			}
 		})
 		Context("Spoke-Hub-Spoke", func() {
@@ -196,7 +173,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineSetResourcePolicy{},
-				Spoke:  &vmopv1a3.VirtualMachineSetResourcePolicy{},
+				Spoke:  &vmopv1a2.VirtualMachineSetResourcePolicy{},
 			}
 		})
 		Context("Spoke-Hub-Spoke", func() {
@@ -216,7 +193,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineWebConsoleRequest{},
-				Spoke:  &vmopv1a3.VirtualMachineWebConsoleRequest{},
+				Spoke:  &vmopv1a2.VirtualMachineWebConsoleRequest{},
 			}
 		})
 		Context("Spoke-Hub-Spoke", func() {
@@ -234,7 +211,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 
 func overrideVirtualMachineFieldsFuncs(codecs runtimeserializer.CodecFactory) []interface{} {
 	return []interface{}{
-		func(vmSpec *vmopv1a3.VirtualMachineSpec, c randfill.Continue) {
+		func(vmSpec *vmopv1a2.VirtualMachineSpec, c randfill.Continue) {
 			c.Fill(vmSpec)
 
 			if bs := vmSpec.Bootstrap; bs != nil {
@@ -263,8 +240,11 @@ func overrideVirtualMachineFieldsFuncs(codecs runtimeserializer.CodecFactory) []
 				}
 			}
 		},
-		func(vmStatus *vmopv1a3.VirtualMachineStatus, c randfill.Continue) {
+		func(vmStatus *vmopv1a2.VirtualMachineStatus, c randfill.Continue) {
 			c.Fill(vmStatus)
+
+			// This field was removed in v1a3.
+			vmStatus.Image = nil
 		},
 		func(msg *json.RawMessage, c randfill.Continue) {
 			*msg = []byte(`{"foo": "bar"}`)
@@ -280,20 +260,6 @@ func overrideVirtualMachineImageFieldsFuncs(codecs runtimeserializer.CodecFactor
 			// Since only VMOP updates the CVMI/VMI's we didn't bother with conversion
 			// when adding this field.
 			vmiStatus.Disks = nil
-		},
-	}
-}
-
-func overrideVirtualMachineImageCacheFieldsFuncs(codecs runtimeserializer.CodecFactory) []interface{} {
-	return []interface{}{
-		func(status *vmopv1.VirtualMachineImageCacheStatus, c randfill.Continue) {
-			c.Fill(status)
-
-			for i := range status.Locations {
-				for j := range status.Locations[i].Files {
-					status.Locations[i].Files[j].Type = ""
-				}
-			}
 		},
 	}
 }
